# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: PermsCraft REST API
  description: |
    Embedded REST API for the PermsCraft Minecraft permission plugin.

    ## Authentication
    Every request must include a valid API key in the `Authorization` header:
    ```
    Authorization: Bearer <api-key>
    ```
    Keys are managed in-game or from console:
    ```
    /pc apikey create <label> [scopes...]
    /pc apikey list
    /pc apikey revoke <label>
    ```

    ## Scopes
    | Scope    | Description |
    |----------|-------------|
    | `read`   | GET endpoints for groups, users, server info |
    | `write`  | POST/PUT/DELETE for groups and users |
    | `log`    | Access to audit log endpoints |
    | `sync`   | Trigger cache flush and sync operations |
    | `backup` | Export, import, and restore backups |
    | `admin`  | All scopes combined |

    ## Security
    > ⚠️ The API transmits keys over plain HTTP by default.
    > It is strongly recommended to use a TLS reverse proxy (nginx/Caddy)
    > and bind to `127.0.0.1` in `config.yml`.
    > See the [TLS setup guide](https://github.com/your-repo/wiki/REST-API-TLS).

    ## Rate Limiting
    Default: 120 requests per minute per IP. Configurable in `config.yml`.
  version: 1.0.0
  contact:
    name: PermsCraft
    url: https://github.com/your-repo
servers:
  - url: http://localhost:4567/api/v1
    description: Local server (default)
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Group 'admin' not found."
    Group:
      type: object
      properties:
        name:
          type: string
          example: admin
        weight:
          type: integer
          example: 100
        prefix:
          type: string
          example: "&c[Admin] "
        suffix:
          type: string
          example: ""
        parents:
          type: array
          items:
            type: string
          example: ["default", "vip"]
        permissions:
          type: array
          items:
            type: string
          example: ["essentials.fly", "-essentials.ban"]
        meta:
          type: object
          additionalProperties:
            type: string
...
paths:
...
  /groups:
    get:
      summary: List all groups
      tags: [Groups]
      responses:
        "200":
          description: List of groups
          content:
            application/json:
              schema:
                type: object
                properties:
                  groups:
                    type: array
                    items:
                      $ref: '#/components/schemas/Group'
                  count:
                    type: integer
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: Create a group
      tags: [Groups]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  example: vip
                weight:
                  type: integer
                  default: 0
                  example: 10
      responses:
        "201":
          description: Group created
        "400":
          description: Invalid input or group already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /groups/{name}:
    get:
      summary: Get a single group
      tags: [Groups]
      parameters:
        - in: path
          name: name
          required: true
          schema:
            type: string
          example: admin
      responses:
        "200":
          description: Group detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        "404":
          description: Group not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Delete a group
      tags: [Groups]
      parameters:
        - in: path
          name: name
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Group deleted
        "404":
          description: Group not found
  /groups/{name}/permissions:
    post:
      summary: Add a permission to a group
      tags: [Groups]
      parameters:
        - in: path
          name: name
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [permission]
              properties:
                permission:
                  type: string
                  example: "essentials.fly"
      responses:
        "200":
          description: Permission added
  /groups/{name}/permissions/{perm}:
    delete:
      summary: Remove a permission from a group
      tags: [Groups]
      parameters:
        - in: path
          name: name
          required: true
          schema:
            type: string
        - in: path
          name: perm
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Permission removed
  /groups/{name}/parents:
    post:
      summary: Add a parent group (inheritance)
      tags: [Groups]
      parameters:
        - in: path
          name: name
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [parent]
              properties:
                parent:
                  type: string
                  example: default
      responses:
        "200":
          description: Parent added
  /groups/{name}/parents/{parent}:
    delete:
      summary: Remove a parent group
      tags: [Groups]
      parameters:
        - in: path
          name: name
          required: true
          schema:
            type: string
        - in: path
          name: parent
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Parent removed
  /groups/{name}/meta:
    put:
      summary: Set a meta key on a group
      tags: [Groups]
      parameters:
        - in: path
          name: name
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [key, value]
              properties:
                key:
                  type: string
                value:
                  type: string
      responses:
        "200":
          description: Meta set
  /groups/{name}/meta/{key}:
    delete:
      summary: Delete a meta key from a group
      tags: [Groups]
      parameters:
        - in: path
          name: name
          required: true
          schema:
            type: string
        - in: path
          name: key
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Meta key deleted
  /groups/{name}/weight:
    put:
      summary: Set group weight
      tags: [Groups]
      parameters:
        - in: path
          name: name
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [weight]
              properties:
                weight:
                  type: integer
                  example: 50
      responses:
        "200":
          description: Weight updated
  /groups/{name}/prefix:
    put:
      summary: Set group chat prefix
      tags: [Groups]
      parameters:
        - in: path
          name: name
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                prefix:
                  type: string
                  example: "&c[Admin] "
      responses:
        "200":
          description: Prefix updated
  /groups/{name}/suffix:
    put:
      summary: Set group chat suffix
      tags: [Groups]
      parameters:
        - in: path
          name: name
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                suffix:
                  type: string
      responses:
        "200":
          description: Suffix updated

  # ── Users ───────────────────────────────────────────────────────────────────
...
  /users/{uuid}/groups:
    post:
      summary: Add user to a group
      tags: [Users]
      parameters:
        - in: path
          name: uuid
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [group]
              properties:
                group:
                  type: string
      responses:
        "200":
          description: User added to group
  /users/{uuid}/groups/{group}:
    delete:
      summary: Remove user from a group
      tags: [Users]
      parameters:
        - in: path
          name: uuid
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: group
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User removed from group
...
  /sync/group/{name}:
    post:
      summary: Sync a specific group across servers
      tags: [Sync]
      parameters:
        - in: path
          name: name
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Group sync triggered
...
  /cache/flush/group/{name}:
    post:
      summary: Flush cache for a specific group
      tags: [Sync]
      parameters:
        - in: path
          name: name
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Group cache flushed

  # ── Backup ──────────────────────────────────────────────────────────────────
...